requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-metadata==4.6.3
lxml==4.6.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
zipp==3.5.0
